+++ backend/src/main/kotlin/com/otoki/powersales/inspection/repository/SiteActivityRepositoryCustomImpl.kt
@@ -22,7 +22,7 @@ class SiteActivityRepositoryCustomImpl(
 
     override fun searchByEmployee(
         employeeId: Long,
-        accountId: Int?,
+        accountId: Long?,
         category: InspectionCategory?,
         fromDate: LocalDate,
         toDate: LocalDate
