+++ backend/src/main/kotlin/com/otoki/powersales/sales/dto/response/MonthlySalesDashboardDetailResponse.kt
@@ -6,7 +6,7 @@ package com.otoki.powersales.sales.dto.response
  * 모바일 월매출조회 페이지의 화면 영역 6종 (진도율 바 / 목표·실적 / 카테고리 4종 / 전년 동월 차트 / 전년 평균 차트) 과 동일한 데이터 구조.
  */
 data class MonthlySalesDashboardDetailResponse(
-    val customerId: Int,
+    val customerId: Long,
     val customerName: String?,
     val salesYear: Int,
     val salesMonth: Int,
